add virtual Keyboard
