+++ common.h
