+++ fire_alarm.cpp
