+++ fire_alarm_init.cpp
