+++ main.cpp
@@ -8,6 +8,7 @@
 
 int main(int argc, char *argv[])
 {
+    qputenv("QT_IM_MODULE", QByteArray("qtvirtualkeyboard"));
     QApplication a(argc, argv);
     QFile file(":/qdarkstyle/dark/darkstyle.qss");
     file.open(QFile::ReadOnly | QFile::Text);
