+++ module_sim_handle.cpp
