+++ password_panel.cpp
